Changed int length to size_t length, but it still does not remove the warning message.  It appears the warning is related to the first parameter (the string) not the result of the strlen.
git-svn-id: https://ctags.svn.sourceforge.net/svnroot/ctags/trunk@484 c5d04d22-be80-434c-894e-aa346cc9e8e8

+++ vim.c
@@ -77,8 +77,8 @@ static jmp_buf Exception;
 #endif
 
 /*
-*	FUNCTION DEFINITIONS
-*/
+ *	FUNCTION DEFINITIONS
+ */
 
 /* This function takes a char pointer, tries to find a scope separator in the
  * string, and if it does, returns a pointer to the character after the colon,
@@ -89,7 +89,7 @@ static const unsigned char* skipPrefix (const unsigned char* name, int *scope)
 {
 	const unsigned char* result = name;
 	int counter;
-	int length;
+	size_t length;
 	length = strlen(name);
 	if (scope != NULL)
 		*scope = '\0';
